+++ Sources/DeckKit/_Deprecated/DeckViewConfiguration+Deprecated.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+
+public extension DeckViewConfiguration {
+
+    @available(*, deprecated, message: "This value will be removed in the next major update. Just create a custom configuration when you need to.")
+    static var down = DeckViewConfiguration(
+        direction: .down
+    )
+}