+++ RELEASE_NOTES.md
@@ -6,6 +6,17 @@ Until then, deprecated features may be removed in the next minor version.
 
 
 
+## 1.0.1
+
+This patch adds support for environment keys.
+
+### 💡 New features
+
+* `DeckViewConfiguration.Key` is a new environment key.
+* `View` has a `.deckViewConfiguration` modifier that you can use to apply a custom view configuration.
+
+
+
 ## 1.0
 
 DeckKit now targets iOS 15.0, macOS 11.0, tvOS 15.0 and watchOS 8.0.